configure JVM args in roborunner.properties

+++ scripts/setup.sh
@@ -6,6 +6,15 @@ if [ ! $# == 2 ]; then
   echo "subdirectories. A thread will be spawned for each instance."
   echo ""
 else
+  if [ -e ./roborunner.properties ]; then
+    echo "WARNING: This will overwrite whatever you have in roborunner.properties."
+    echo "Continue? [Y/n]"
+    read -s -n 1 CONTINUE
+    if [ "$CONTINUE" = 'n' -o "$CONTINUE" = 'N' ]; then
+      echo "Exiting..."
+      exit
+    fi
+  fi
   if [ -e ./robocodes ]; then
     echo "Clearing existing Robocode installs..."
     rm -rf ./robocodes/*
@@ -38,5 +47,6 @@ else
     echo "  Done!"
   fi
   echo "robocodePaths=$paths" > roborunner.properties
+  echo "jvmArgs=-Xmx512M" >> roborunner.properties
 fi
